Separate out GraphqlMutableStoreBase
