+++ packages/webapp/src/stores/graphqlStoreBase.ts
@@ -1,4 +1,4 @@
-import { action, flow, observable } from "mobx";
+import { flow, observable } from "mobx";
 import { ApolloQueryResult } from "apollo-client";
 import { DocumentNode } from "graphql";
 
@@ -6,16 +6,16 @@ import ApolloClient from "../services/ApolloClient";
 
 import StoreBase from "./StoreBase";
 
-interface IdType {
+export interface IdType {
   id: string;
   __typename?: string;
 }
 
-interface QueryResultExtractor<TResult, TQuery> {
+export interface QueryResultExtractor<TResult, TQuery> {
   (queryData: TQuery): TResult[];
 }
 
-interface QueryResultPatcher<TResult> {
+export interface QueryResultPatcher<TResult> {
   (data: TResult): TResult;
 }
 
@@ -73,18 +73,4 @@ export default class GraphqlStoreBase<T extends IdType, TQuery> extends StoreBas
       this.state = "error";
     }
   });
-
-  @action
-  updateItem(item: T) {
-    const updatedItemIndex = this.data.findIndex(existingItem => existingItem.id === item.id);
-
-    // Remove GraphQL-injected fields that won't be accepted in a GraphQL update
-    if (item.__typename) {
-      delete item.__typename;
-    }
-
-    // TODO: Persist change
-
-    this.data[updatedItemIndex] = item;
-  }
 }
